Move operations and operands class to subfolders
