+++ .gitignore
@@ -5,3 +5,4 @@
 *.DS_Store
 .idea/*
 target/*
+/dependency-reduced-pom.xml
