+++ src/main/java/calculator/operand/Operand.java
@@ -1,4 +1,6 @@
-package calculator;
+package calculator.operand;
+
+import calculator.Value;
 
 public interface Operand<T> {
     Value<T> plus(Value<T> other);
