+++ src/test/resources/junit-platform.properties
@@ -1,1 +1,1 @@
-cucumber.publish.enabled=true
+#cucumber.publish.enabled=true
